hciattach: Fix scheduling alarm at end of rtk_init
The use of h5_tshy_sig_alarm is intended to be from a SIGALARM event handler.
For some reason, it has been used in a normal fonction callback, and it is
now trigerring a signal. That signal will trigger a crash later by calling the
handler that was registered. Let's remove the call to the signal handler and
add a function that does the same thing but do not trigger signal.

Tracked-On: OAM-1302
Change-Id: Ice5716c68c6eccfb5dd774a75d7c0f1e043b566d
Signed-off-by: Frédéric Dalleau <frederic.dalleau@linux.intel.com>
Reviewed-on: 419735